requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
colorama>=0.4.6
pydantic-ai>=0.0.14
